+++ common/visualization.py
@@ -15,209 +15,173 @@ import numpy as np
 import subprocess as sp
 
 def get_resolution(filename):
-	command = ['ffprobe', '-v', 'error', '-select_streams', 'v:0',
-			   '-show_entries', 'stream=width,height', '-of', 'csv=p=0', filename]
-	with sp.Popen(command, stdout=sp.PIPE, bufsize=-1) as pipe:
-		for line in pipe.stdout:
-			w, h = line.decode().strip().split(',')
-			return int(w), int(h)
+    command = ['ffprobe', '-v', 'error', '-select_streams', 'v:0',
+               '-show_entries', 'stream=width,height', '-of', 'csv=p=0', filename]
+    with sp.Popen(command, stdout=sp.PIPE, bufsize=-1) as pipe:
+        for line in pipe.stdout:
+            w, h = line.decode().strip().split(',')
+            return int(w), int(h)
 
 def read_video(filename, skip=0, limit=-1):
-	w, h = get_resolution(filename)
-	
-	command = ['ffmpeg',
-			'-i', filename,
-			'-f', 'image2pipe',
-			'-pix_fmt', 'rgb24',
-			'-vsync', '0',
-			'-vcodec', 'rawvideo', '-']
-	
-	i = 0
-	with sp.Popen(command, stdout = sp.PIPE, bufsize=-1) as pipe:
-		while True:
-			data = pipe.stdout.read(w*h*3)
-			if not data:
-				break
-			i += 1
-			if i > skip:
-				yield np.frombuffer(data, dtype='uint8').reshape((h, w, 3))
-			if i == limit:
-				break
-				
-				
-	
+    w, h = get_resolution(filename)
+    
+    command = ['ffmpeg',
+            '-i', filename,
+            '-f', 'image2pipe',
+            '-pix_fmt', 'rgb24',
+            '-vsync', '0',
+            '-vcodec', 'rawvideo', '-']
+    
+    i = 0
+    with sp.Popen(command, stdout = sp.PIPE, bufsize=-1) as pipe:
+        while True:
+            data = pipe.stdout.read(w*h*3)
+            if not data:
+                break
+            i += 1
+            if i > skip:
+                yield np.frombuffer(data, dtype='uint8').reshape((h, w, 3))
+            if i == limit:
+                break
+                
+                
+    
 def downsample_tensor(X, factor):
-	length = X.shape[0]//factor * factor
-	return np.mean(X[:length].reshape(-1, factor, *X.shape[1:]), axis=1)
+    length = X.shape[0]//factor * factor
+    return np.mean(X[:length].reshape(-1, factor, *X.shape[1:]), axis=1)
 
 def render_animation(keypoints, poses, skeleton, fps, bitrate, azim, output, viewport,
-					 limit=-1, downsample=1, size=6, input_video_path=None, input_video_skip=0):
-	"""
-	TODO
-	Render an animation. The supported output modes are:
-	 -- 'interactive': display an interactive figure
-					   (also works on notebooks if associated with %matplotlib inline)
-	 -- 'html': render the animation as HTML5 video. Can be displayed in a notebook using HTML(...).
-	 -- 'filename.mp4': render and export the animation as an h264 video (requires ffmpeg).
-	 -- 'filename.gif': render and export the animation a gif file (requires imagemagick).
-	"""
-	plt.ioff()
-	fig = plt.figure(figsize=(size*(1 + len(poses)), size))
-	ax_in = fig.add_subplot(1, 1 + len(poses), 1)
-	ax_in.get_xaxis().set_visible(False)
-	ax_in.get_yaxis().set_visible(False)
-	ax_in.set_axis_off()
-	ax_in.set_title('Input')
+                     limit=-1, downsample=1, size=6, input_video_path=None, input_video_skip=0):
+    """
+    TODO
+    Render an animation. The supported output modes are:
+     -- 'interactive': display an interactive figure
+                       (also works on notebooks if associated with %matplotlib inline)
+     -- 'html': render the animation as HTML5 video. Can be displayed in a notebook using HTML(...).
+     -- 'filename.mp4': render and export the animation as an h264 video (requires ffmpeg).
+     -- 'filename.gif': render and export the animation a gif file (requires imagemagick).
+    """
+    plt.ioff()
+    fig = plt.figure(figsize=(size*(1 + len(poses)), size))
+    ax_in = fig.add_subplot(1, 1 + len(poses), 1)
+    ax_in.get_xaxis().set_visible(False)
+    ax_in.get_yaxis().set_visible(False)
+    ax_in.set_axis_off()
+    ax_in.set_title('Input')
 
-	ax_3d = []
-	lines_3d = []
-	trajectories = []
-	radius = 1.7
-	for index, (title, data) in enumerate(poses.items()):
-		ax = fig.add_subplot(1, 1 + len(poses), index+2, projection='3d')
-		ax.view_init(elev=15., azim=azim)
-		ax.set_xlim3d([-radius/2, radius/2])
-		ax.set_zlim3d([0, radius])
-		ax.set_ylim3d([-radius/2, radius/2])
-		ax.set_aspect('equal')
-		ax.set_xticklabels([])
-		ax.set_yticklabels([])
-		ax.set_zticklabels([])
-		ax.dist = 7.5
-		ax.set_title(title) #, pad=35
-		ax_3d.append(ax)
-		lines_3d.append([])
-		trajectories.append(data[:, 0, [0, 1]])
-	poses = list(poses.values())
+    ax_3d = []
+    lines_3d = []
+    trajectories = []
+    radius = 1.7
+    for index, (title, data) in enumerate(poses.items()):
+        ax = fig.add_subplot(1, 1 + len(poses), index+2, projection='3d')
+        ax.view_init(elev=15., azim=azim)
+        ax.set_xlim3d([-radius/2, radius/2])
+        ax.set_zlim3d([0, radius])
+        ax.set_ylim3d([-radius/2, radius/2])
+        ax.set_aspect('equal')
+        ax.set_xticklabels([])
+        ax.set_yticklabels([])
+        ax.set_zticklabels([])
+        ax.dist = 12.5
+        ax.set_title(title) #, pad=35
+        ax_3d.append(ax)
+        lines_3d.append([])
+        trajectories.append(data[:, 0, [0, 1]])
+    poses = list(poses.values())
 
-	# Decode video
-	if input_video_path is None:
-		# Black background
-		all_frames = np.zeros((keypoints.shape[0], viewport[1], viewport[0]), dtype='uint8')
-	else:
-		# Load video using ffmpeg
-		all_frames = []
-		for f in read_video(input_video_path, skip=input_video_skip):
-			all_frames.append(f)
-		effective_length = min(keypoints.shape[0], len(all_frames))
-		all_frames = all_frames[:effective_length]
-		
-	if downsample > 1:
-		keypoints = downsample_tensor(keypoints, downsample)
-		all_frames = downsample_tensor(np.array(all_frames), downsample).astype('uint8')
-		for idx in range(len(poses)):
-			poses[idx] = downsample_tensor(poses[idx], downsample)
-			trajectories[idx] = downsample_tensor(trajectories[idx], downsample)
-		fps /= downsample
+    # Decode video
+    if input_video_path is None:
+        # Black background
+        all_frames = np.zeros((keypoints.shape[0], viewport[1], viewport[0]), dtype='uint8')
+    else:
+        # Load video using ffmpeg
+        all_frames = []
+        for f in read_video(input_video_path, skip=input_video_skip):
+            all_frames.append(f)
+        effective_length = min(keypoints.shape[0], len(all_frames))
+        all_frames = all_frames[:effective_length]
+    
+    if downsample > 1:
+        keypoints = downsample_tensor(keypoints, downsample)
+        all_frames = downsample_tensor(np.array(all_frames), downsample).astype('uint8')
+        for idx in range(len(poses)):
+            poses[idx] = downsample_tensor(poses[idx], downsample)
+            trajectories[idx] = downsample_tensor(trajectories[idx], downsample)
+        fps /= downsample
 
-	initialized = False
-	image = None
-	lines = []
-	points = None
-	
-	if limit < 1:
-		limit = len(all_frames)
-	else:
-		limit = min(limit, len(all_frames))
+    initialized = False
+    image = None
+    lines = []
+    points = None
+    
+    if limit < 1:
+        limit = len(all_frames)
+    else:
+        limit = min(limit, len(all_frames))
 
-	parents = skeleton.parents()
-	def update_video(i):
-		nonlocal initialized, image, lines, points
+    parents = skeleton.parents()
+    def update_video(i):
+        nonlocal initialized, image, lines, points
 
-		for n, ax in enumerate(ax_3d):
-			ax.set_xlim3d([-radius/2 + trajectories[n][i, 0], radius/2 + trajectories[n][i, 0]])
-			ax.set_ylim3d([-radius/2 + trajectories[n][i, 1], radius/2 + trajectories[n][i, 1]])
+        for n, ax in enumerate(ax_3d):
+            ax.set_xlim3d([-radius/2 + trajectories[n][i, 0], radius/2 + trajectories[n][i, 0]])
+            ax.set_ylim3d([-radius/2 + trajectories[n][i, 1], radius/2 + trajectories[n][i, 1]])
 
+        # Update 2D poses
+        if not initialized:
+            image = ax_in.imshow(all_frames[i], aspect='equal')
+            
+            for j, j_parent in enumerate(parents):
+                if j_parent == -1:
+                    continue
+                    
+                if len(parents) == keypoints.shape[1] and 1 == 2:
+                    # Draw skeleton only if keypoints match (otherwise we don't have the parents definition)
+                    lines.append(ax_in.plot([keypoints[i, j, 0], keypoints[i, j_parent, 0]],
+                                            [keypoints[i, j, 1], keypoints[i, j_parent, 1]], color='pink'))
 
-		# the coco_keypoints are used here and its not the same as the 3D skeleton model.
-		coco_parents = [
-			'nose',
-			'left_eye',
-			'right_eye',
-			'left_ear',
-			'right_ear',
-			'left_shoulder',
-			'right_shoulder',
-			'left_elbow',
-			'right_elbow',
-			'left_wrist',
-			'right_wrist',
-			'left_hip',
-			'right_hip',
-			'left_knee',
-			'right_knee',
-			'left_ankle',
-			'right_ankle'
-		]
-		#parents_coco_head = [1, 2, 0, 1, 2,]
-		#parents_coco_torso = [20,20,5,6,7,8]
-		#parents_coco_bottom = [21,21,11,12,15,16, 0, 17]
-		parents_coco = [1, 2, 0, 1, 2, 17,17,5,6,7,8, 18, 18,11,12,13 ,14, 0, 17]
-		# This is needed because neck and hip does not appear naturally as points in the coco keypoint representation
-		neck = np.zeros((1,2))
-		neck[0][0] = (keypoints[i, 5, 0]+ keypoints[i, 6, 0])/2
-		neck[0][1] = (keypoints[i, 5, 1] + keypoints[i, 6, 1]) / 2
-		keypoints_plus_c_hip_and_neck = np.append(keypoints[i], neck, axis=0)
-		hip = np.zeros((1,2))
-		hip[0][0] = (keypoints[i, 11, 0] + keypoints[i, 12, 0])/2
-		hip[0][1] = (keypoints[i, 11, 1] + keypoints[i, 12, 1])/2
-		keypoints_plus_c_hip_and_neck = np.append(keypoints_plus_c_hip_and_neck, hip, axis=0)
-		# Update 2D poses
-		if not initialized:
-			image = ax_in.imshow(all_frames[i], aspect='equal')
+                col = 'red' if j in skeleton.joints_right() else 'black'
+                for n, ax in enumerate(ax_3d):
+                    pos = poses[n][i]
+                    lines_3d[n].append(ax.plot([pos[j, 0], pos[j_parent, 0]],
+                                               [pos[j, 1], pos[j_parent, 1]],
+                                               [pos[j, 2], pos[j_parent, 2]], zdir='z', c=col))
 
-			for j, j_parent in enumerate(parents_coco):
-				if j < 17 and parents[j] == -1:
-					continue
+            points = ax_in.scatter(*keypoints[i].T, 5, color='red', edgecolors='white', zorder=10)
 
-				if len(parents_coco) == keypoints_plus_c_hip_and_neck.shape[0]:
-					# Draw skeleton only if keypoints match (otherwise we don't have the parents definition)
-					lines.append(ax_in.plot([keypoints_plus_c_hip_and_neck[j, 0], keypoints_plus_c_hip_and_neck[j_parent, 0]],
-											[keypoints_plus_c_hip_and_neck[j, 1], keypoints_plus_c_hip_and_neck[j_parent, 1]], color='pink'))
+            initialized = True
+        else:
+            image.set_data(all_frames[i])
 
-				col = 'red' if j in skeleton.joints_right() else 'black'
-				if j < 17:
-					for n, ax in enumerate(ax_3d):
-						pos = poses[n][i]
-						lines_3d[n].append(ax.plot([pos[j, 0], pos[parents[j], 0]],
-													[pos[j, 1], pos[parents[j], 1]],
-													[pos[j, 2], pos[parents[j], 2]], zdir='z', c=col))
+            for j, j_parent in enumerate(parents):
+                if j_parent == -1:
+                    continue
+                
+                if len(parents) == keypoints.shape[1] and 1 == 2:
+                    lines[j-1][0].set_data([keypoints[i, j, 0], keypoints[i, j_parent, 0]],
+                                           [keypoints[i, j, 1], keypoints[i, j_parent, 1]])
 
-			points = ax_in.scatter(*keypoints[i].T, 5, color='red', edgecolors='white', zorder=10)
+                for n, ax in enumerate(ax_3d):
+                    pos = poses[n][i]
+                    lines_3d[n][j-1][0].set_xdata([pos[j, 0], pos[j_parent, 0]])
+                    lines_3d[n][j-1][0].set_ydata([pos[j, 1], pos[j_parent, 1]])
+                    lines_3d[n][j-1][0].set_3d_properties([pos[j, 2], pos[j_parent, 2]], zdir='z')
 
-			initialized = True
-		else:
-			image.set_data(all_frames[i])
+            points.set_offsets(keypoints[i])
+        
+        print('{}/{}      '.format(i, limit), end='\r')
+        
 
-			for j, j_parent in enumerate(parents_coco):
-				if j < 17 and parents[j] == -1:
-					continue
-				
-				if len(parents_coco) == keypoints_plus_c_hip_and_neck.shape[0]:
-					lines[j-1][0].set_data([keypoints_plus_c_hip_and_neck[j, 0], keypoints_plus_c_hip_and_neck[j_parent, 0]],
-											[keypoints_plus_c_hip_and_neck[j, 1], keypoints_plus_c_hip_and_neck[j_parent, 1]])
-				if j < 17:
-					for n, ax in enumerate(ax_3d):
-						pos = poses[n][i]
-						lines_3d[n][j-1][0].set_xdata([pos[j, 0], pos[parents[j], 0]])
-						lines_3d[n][j-1][0].set_ydata([pos[j, 1], pos[parents[j], 1]])
-						lines_3d[n][j-1][0].set_3d_properties([pos[j, 2], pos[parents[j], 2]], zdir='z')
-
-			points.set_offsets(keypoints[i])
-		
-		print('{}/{}      '.format(i, limit), end='\r')
-		
-
-	#fig.tight_layout()
-	
-	anim = FuncAnimation(fig, update_video, frames=np.arange(0, limit), interval=1000/fps, repeat=False)
-	
-	if output.endswith('.mp4'):
-		Writer = writers['ffmpeg']
-		writer = Writer(fps=fps, metadata={}, bitrate=bitrate)
-		anim.save(output, writer=writer)
-	elif output.endswith('.gif'):
-		anim.save(output, dpi=80, writer='imagemagick')
-	else:
-		raise ValueError('Unsupported output format (only .mp4 and .gif are supported)')
-	plt.close()
+    fig.tight_layout()
+    
+    anim = FuncAnimation(fig, update_video, frames=np.arange(0, limit), interval=1000/fps, repeat=False)
+    if output.endswith('.mp4'):
+        Writer = writers['ffmpeg']
+        writer = Writer(fps=fps, metadata={}, bitrate=bitrate)
+        anim.save(output, writer=writer)
+    elif output.endswith('.gif'):
+        anim.save(output, dpi=80, writer='imagemagick')
+    else:
+        raise ValueError('Unsupported output format (only .mp4 and .gif are supported)')
+    plt.close()
