--- data/tools/Notes_export_detectron
@@ -1,8 +1,0 @@
-   # Add those lines to the end of your main function in the infer_simple detectron file to export the 2D pose data from images for the 3D detection.
-
-    dictionarry_keypoints={'S1': {'Directions' : np.asarray([glob_keypoints])}}
-    metadata = {'layout_name': 'h36m', 'num_joints': 17, 'keypoints_symmetry': [[4, 5, 6, 11, 12, 13], [1, 2, 3, 14, 15, 16]]}
-    #np.savez(os.path.join(args.output_dir, "detections.npz"), metadata = metadata, positions_2d=dictionarry_keypoints)
-    np.savez(os.path.join('/home/narvis/Dev/VideoPose3D/data', "data_2d_h36m_detections.npz"), metadata=metadata, positions_2d=dictionarry_keypoints)
-
-
