+++ detectron_tools/infer_simple.py
@@ -190,11 +190,11 @@ def main(args):
         pose_x_y_prob_after_softmax = cls_keyps_max_prob[[0,1,3]]
         glob_keypoints.append(np.transpose(pose_x_y_prob_after_softmax))
 
-    dictionarry_keypoints={'S1': {'Directions' : np.asarray([glob_keypoints])}}
+    dictionarry_keypoints={'S1': {'Directions 1' : np.asarray([glob_keypoints])}}
     metadata = {'layout_name': 'h36m', 'num_joints': 17, 'keypoints_symmetry': [[4, 5, 6, 11, 12, 13], [1, 2, 3, 14, 15, 16]]}
-    #np.savez(os.path.join(args.output_dir, "detections.npz"), metadata = metadata, positions_2d=dictionarry_keypoints)
-    np.savez(os.path.join('pathToSave/', "data_2d_detections.npz"), metadata=metadata, positions_2d=dictionarry_keypoints)
-    #Splitting Command:  ffmpeg -i ice.mp4 -r 25/1 splitted_scating/output%04d.png
+    np.savez(os.path.join('/home/narvis/Dev/VideoPose3D/data', "data_2d_detections.npz"), metadata=metadata, positions_2d=dictionarry_keypoints)
+    np.savez(os.path.join(args.output_dir, "data_2d_detections.npz"), metadata=metadata, positions_2d=dictionarry_keypoints)
+    #Splitting Command:  ffmpeg -i ice_cutted_cropped.mp4 -r 25/1 splitted_scating/output%04d.png
     # Video Link: https://www.youtube.com/watch?v=ke0iusvydl8
