--- inTheWild.py
@@ -1,57 +1,0 @@
-import numpy as np
-
-from common.arguments import parse_args
-import torch
-
-import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-import os
-import sys
-import errno
-
-from common.camera import *
-from common.model import *
-from common.loss import *
-from common.generators import ChunkedGenerator, UnchunkedGenerator
-from time import time
-from common.utils import deterministic_random
-
-
-print('Loading 2D detections...')
-keypoints = np.load('data/data_2d_' + args.dataset + '_' + args.keypoints + '.npz')
-keypoints_symmetry = keypoints['metadata'].item()['keypoints_symmetry']
-kps_left, kps_right = list(keypoints_symmetry[0]), list(keypoints_symmetry[1])
-joints_left, joints_right = list(dataset.skeleton().joints_left()), list(dataset.skeleton().joints_right())
-keypoints = keypoints['positions_2d'].item()
-
-
-def predict(test_generator):
-    model_pos.eval()
-    for _, batch, batch_2d in test_generator.next_epoch():
-        inputs_2d = torch.from_numpy(batch_2d.astype('float32'))
-        if torch.cuda.is_available():
-            inputs_2d = inputs_2d.cuda()
-
-        # Positional model
-        predicted_3d_pos = model_pos(inputs_2d)
-
-        # Test-time augmentation (if enabled)
-        if test_generator.augment_enabled():
-            # Undo flipping and take average with non-flipped version
-            predicted_3d_pos[1, :, :, 0] *= -1
-            predicted_3d_pos[1, :, joints_left + joints_right] = predicted_3d_pos[1, :, joints_right + joints_left]
-            predicted_3d_pos = torch.mean(predicted_3d_pos, dim=0, keepdim=True)
-
-        if return_predictions:
-            return predicted_3d_pos.squeeze(0).cpu().numpy()
-
-def inTheWild():
-    gen = UnchunkedGenerator(None, None, [input_keypoints],
-                             pad=pad, causal_shift=causal_shift, augment=args.test_time_augmentation,
-                             kps_left=kps_left, kps_right=kps_right, joints_left=joints_left, joints_right=joints_right)
-    prediction = predict(gen)
-
-
-if __name__ == '__main__':
-    inTheWild()
