+++ run_wild.py
@@ -4,204 +4,385 @@
 # This source code is licensed under the license found in the
 # LICENSE file in the root directory of this source tree.
 #
-#
-# How to call this file: -k detections -arc 3,3,3,3,3 -c checkpoint --evaluate d-pt-243.bin --render --viz-subject S1 --viz-action Directions --viz-camera 0 --viz-output output_scater.mp4 --viz-size 5 --viz-downsample 1 --viz-video /pathToVideo/ice_cutted_cropped.mp4 --viz-skip 0 --viz-limit 240
+
+# Run arguments: python run_wild -k detections -arc 3,3,3,3,3 -c checkpoint --evaluate d-pt-243.bin --render --viz-subject S1 --viz-action Directions
+# --viz-video /home/narvis/Dev/scater_1080/out_cutted.mp4 --viz-camera 0 --viz-output output_scater.mp4 --viz-size 5 --viz-downsample 1 --viz-skip 9
 
 import numpy as np
 
 from common.arguments import parse_args
 import torch
 
+import torch.nn as nn
+import torch.nn.functional as F
+import torch.optim as optim
 import os
+import sys
 import errno
 
 from common.camera import *
 from common.model import *
+from common.loss import *
 from common.generators import ChunkedGenerator, UnchunkedGenerator
+from time import time
 from common.utils import deterministic_random
-from common.skeleton import Skeleton
-
-from PIL import Image, ImageDraw
 
 args = parse_args()
 print(args)
 
 try:
-	# Create checkpoint directory if it does not exist
-	os.makedirs(args.checkpoint)
+    # Create checkpoint directory if it does not exist
+    os.makedirs(args.checkpoint)
 except OSError as e:
-	if e.errno != errno.EEXIST:
-		raise RuntimeError('Unable to create checkpoint directory:', args.checkpoint)
+    if e.errno != errno.EEXIST:
+        raise RuntimeError('Unable to create checkpoint directory:', args.checkpoint)
 
+
+print('Loading dataset...')
+dataset_path = 'data/data_3d_' + args.dataset + '.npz'
+if args.dataset == 'h36m':
+    from common.h36m_dataset import Human36mDataset
+    dataset = Human36mDataset(dataset_path)
+elif args.dataset.startswith('humaneva'):
+    from common.humaneva_dataset import HumanEvaDataset
+    dataset = HumanEvaDataset(dataset_path)
+else:
+    raise KeyError('Invalid dataset')
+
+print('Preparing data...')
+for subject in dataset.subjects():
+    for action in dataset[subject].keys():
+        anim = dataset[subject][action]
+
+        positions_3d = []
+        for cam in anim['cameras']:
+            pos_3d = world_to_camera(anim['positions'], R=cam['orientation'], t=cam['translation'])
+            pos_3d[:, 1:] -= pos_3d[:, :1] # Remove global offset, but keep trajectory in first position
+            positions_3d.append(pos_3d)
+        anim['positions_3d'] = positions_3d
 
 print('Loading 2D detections...')
-# keypoints contains the loaded npz-file with all the 2D positions exported by Detectron in 'positions_2D' and metainfo in 'metadata'
+#keypoints = np.load('data/data_2d_' + args.dataset + '_' + args.keypoints + '.npz')
 keypoints = np.load('data/data_2d_' + args.keypoints + '.npz')
 keypoints_symmetry = keypoints['metadata'].item()['keypoints_symmetry']
 kps_left, kps_right = list(keypoints_symmetry[0]), list(keypoints_symmetry[1])
-joints_left, joints_right = (kps_left, kps_right) #list(dataset.skeleton().joints_left()), list(dataset.skeleton().joints_right())
+kps_left = [1, 3, 5, 7, 9, 11, 13, 15]
+kps_right = [2, 4, 6, 8, 10, 12, 14, 16]
+joints_left, joints_right = list(dataset.skeleton().joints_left()), list(dataset.skeleton().joints_right())
 keypoints = keypoints['positions_2d'].item()
-
-
-## visualize 2d just used if you wanna see if your keypoints 2D are correct
-#r = 4
-#squeezed_kp = np.squeeze(keypoints['S1']['Directions'])
-#for i in range(squeezed_kp.shape[0]):
-#	image = Image.new('RGB', (960, 720))
-#	draw = ImageDraw.Draw(image)
-#	for j in range(squeezed_kp.shape[1]):
-#		draw.ellipse((squeezed_kp[i,j,0]-r, squeezed_kp[i,j,1]-r, squeezed_kp[i,j,0]+r, squeezed_kp[i,j,1]+r), fill=(255,0,0,255))
-#	image.save(f'keypoint_vis/2dkey_{i}.png')
+# keypoints['S1']['Directions 1'][0].shape == 1384,17,2 should be 3 though
+# when loading the file provided py pavllo as example the shape is fine (1384,17,3) and the joints left = <class 'list'>: [4, 5, 6, 11, 12, 13] and right = <class 'list'>: [1, 2, 3, 14, 15, 16]
 
 
 
-for subject in keypoints.keys():
-	for action in keypoints[subject]:
-		for cam_idx, kps in enumerate(keypoints[subject][action]):
-			# Normalize camera frame
-			kps[..., :2] = normalize_screen_coordinates(kps[..., :2], w=960, h=720)
-			keypoints[subject][action][cam_idx] = kps
 
+subject = 'S1'
+
+action = 'Directions'
+
+#for subject in keypoints.keys():
+#    for action in keypoints[subject]:
+#        for cam_idx, kps in enumerate(keypoints[subject][action]):
+#            # Normalize camera frame
+#            cam = dataset.cameras()[subject][cam_idx]
+#            kps[..., :2] = normalize_screen_coordinates(kps[..., :2], w=cam['res_w'], h=cam['res_h'])
+#            keypoints[subject][action][cam_idx] = kps
+
+width_of = 1080
+height_of = 1080
+
+for cam_idx, kps in enumerate(keypoints[subject][action]):
+
+    # Normalize camera frame
+    cam = dataset.cameras()[subject][cam_idx]
+    kps[..., :2] = normalize_screen_coordinates(kps[..., :2], w=width_of, h=height_of)
+    keypoints[subject][action][cam_idx] = kps
+
+subjects_train = args.subjects_train.split(',')
+subjects_semi = [] if not args.subjects_unlabeled else args.subjects_unlabeled.split(',')
 subjects_test = args.subjects_test.split(',')
 
+semi_supervised = len(subjects_semi) > 0
+if semi_supervised and not dataset.supports_semi_supervised():
+    raise RuntimeError('Semi-supervised training is not implemented for this dataset')
+            
 def fetch(subjects, action_filter=None, subset=1, parse_3d_poses=True):
-	out_poses_3d = []
-	out_poses_2d = []
-	out_camera_params = []
-	for subject in subjects:
-		for action in keypoints[subject].keys():
-			if action_filter is not None:
-				found = False
-				for a in action_filter:
-					if action.startswith(a):
-						found = True
-						break
-				if not found:
-					continue
+    out_poses_3d = []
+    out_poses_2d = []
+    out_camera_params = []
+    subject = 'S1'
+    action = 'Directions'
 
-			poses_2d = keypoints[subject][action]
-			for i in range(len(poses_2d)):  # Iterate across cameras
-				out_poses_2d.append(poses_2d[i])
+                
+    poses_2d = keypoints[subject][action]
+    for i in range(len(poses_2d)): # Iterate across cameras
+        out_poses_2d.append(poses_2d[i])
 
-	if len(out_camera_params) == 0:
-		out_camera_params = None
-	if len(out_poses_3d) == 0:
-		out_poses_3d = None
 
-	stride = args.downsample
-	if subset < 1:
-		for i in range(len(out_poses_2d)):
-			n_frames = int(round(len(out_poses_2d[i]) // stride * subset) * stride)
-			start = deterministic_random(0, len(out_poses_2d[i]) - n_frames + 1, str(len(out_poses_2d[i])))
-			out_poses_2d[i] = out_poses_2d[i][start:start + n_frames:stride]
-			if out_poses_3d is not None:
-				out_poses_3d[i] = out_poses_3d[i][start:start + n_frames:stride]
-	elif stride > 1:
-		# Downsample as requested
-		for i in range(len(out_poses_2d)):
-			out_poses_2d[i] = out_poses_2d[i][::stride]
-			if out_poses_3d is not None:
-				out_poses_3d[i] = out_poses_3d[i][::stride]
+    if len(out_camera_params) == 0:
+        out_camera_params = None
+    if len(out_poses_3d) == 0:
+        out_poses_3d = None
+    
+    stride = args.downsample
+    if subset < 1:
+        for i in range(len(out_poses_2d)):
+            n_frames = int(round(len(out_poses_2d[i])//stride * subset)*stride)
+            start = deterministic_random(0, len(out_poses_2d[i]) - n_frames + 1, str(len(out_poses_2d[i])))
+            out_poses_2d[i] = out_poses_2d[i][start:start+n_frames:stride]
+            if out_poses_3d is not None:
+                out_poses_3d[i] = out_poses_3d[i][start:start+n_frames:stride]
+    elif stride > 1:
+        # Downsample as requested
+        for i in range(len(out_poses_2d)):
+            out_poses_2d[i] = out_poses_2d[i][::stride]
+            if out_poses_3d is not None:
+                out_poses_3d[i] = out_poses_3d[i][::stride]
+    
 
-	return out_camera_params, out_poses_3d, out_poses_2d
+    return out_camera_params, out_poses_3d, out_poses_2d
 
-cameras_valid, poses_valid, poses_valid_2d = fetch(subjects_test)
+action_filter = None if args.actions == '*' else args.actions.split(',')
+if action_filter is not None:
+    print('Selected actions:', action_filter)
+    
+cameras_valid, poses_valid, poses_valid_2d = fetch(subjects_test, action_filter)
 
 filter_widths = [int(x) for x in args.architecture.split(',')]
 
+model_pos = TemporalModel(poses_valid_2d[0].shape[-2], poses_valid_2d[0].shape[-1], 17,
+                            filter_widths=filter_widths, causal=args.causal, dropout=args.dropout, channels=args.channels,
+                            dense=args.dense)
 
-model_pos = TemporalModel(poses_valid_2d[0].shape[-2], poses_valid_2d[0].shape[-1], poses_valid_2d[0].shape[-2],
-						  filter_widths=filter_widths, causal=args.causal, dropout=0.0, channels=args.channels,
-						  dense=args.dense)
 
 receptive_field = model_pos.receptive_field()
 print('INFO: Receptive field: {} frames'.format(receptive_field))
-pad = (receptive_field - 1) // 2  # Padding on each side
+pad = (receptive_field - 1) // 2 # Padding on each side
 if args.causal:
-	print('INFO: Using causal convolutions')
-	causal_shift = pad
+    print('INFO: Using causal convolutions')
+    causal_shift = pad
 else:
-	causal_shift = 0
+    causal_shift = 0
 
-# Shows how many parameter Model contains
 model_params = 0
 for parameter in model_pos.parameters():
-	model_params += parameter.numel()
+    model_params += parameter.numel()
 print('INFO: Trainable parameter count:', model_params)
 
 if torch.cuda.is_available():
-	model_pos = model_pos.cuda()
-
+    model_pos = model_pos.cuda()
+#    model_pos_train = model_pos_train.cuda()
+    
 if args.resume or args.evaluate:
-	chk_filename = os.path.join(args.checkpoint, args.resume if args.resume else args.evaluate)
-	print('Loading checkpoint', chk_filename)
-	checkpoint = torch.load(chk_filename, map_location=lambda storage, loc: storage)
-	print('This model was trained for {} epochs'.format(checkpoint['epoch']))
-	model_pos.load_state_dict(checkpoint['model_pos'])
-
+    chk_filename = os.path.join(args.checkpoint, args.resume if args.resume else args.evaluate)
+    print('Loading checkpoint', chk_filename)
+    checkpoint = torch.load(chk_filename, map_location=lambda storage, loc: storage)
+    print('This model was trained for {} epochs'.format(checkpoint['epoch']))
+#    model_pos_train.load_state_dict(checkpoint['model_pos'])
+    model_pos.load_state_dict(checkpoint['model_pos'])
+    
 test_generator = UnchunkedGenerator(cameras_valid, poses_valid, poses_valid_2d,
-									pad=pad, causal_shift=causal_shift, augment=False,
-									kps_left=kps_left, kps_right=kps_right, joints_left=joints_left,
-									joints_right=joints_right)
+                                    pad=pad, causal_shift=causal_shift, augment=False,
+                                    kps_left=kps_left, kps_right=kps_right, joints_left=joints_left, joints_right=joints_right)
 print('INFO: Testing on {} frames'.format(test_generator.num_frames()))
 
-# Evaluate
 
 def evaluate(test_generator, action=None, return_predictions=False):
-	with torch.no_grad():
-		model_pos.eval()
-		N = 0
-		for _, batch, batch_2d in test_generator.next_epoch():
-			inputs_2d = torch.from_numpy(batch_2d.astype('float32'))
-			if torch.cuda.is_available():
-				inputs_2d = inputs_2d.cuda()
+    epoch_loss_3d_pos = 0
+    epoch_loss_3d_pos_procrustes = 0
+    epoch_loss_3d_pos_scale = 0
+    epoch_loss_3d_vel = 0
+    with torch.no_grad():
+        model_pos.eval()
+        N = 0
+        for _, batch, batch_2d in test_generator.next_epoch():
+            inputs_2d = torch.from_numpy(batch_2d.astype('float32'))
+            if torch.cuda.is_available():
+                inputs_2d = inputs_2d.cuda()
 
-			# Positional model
-			predicted_3d_pos = model_pos(inputs_2d)
+            # Positional model
+            predicted_3d_pos = model_pos(inputs_2d)
 
-			# Test-time augmentation (if enabled)
-			if test_generator.augment_enabled():
-				# Undo flipping and take average with non-flipped version
-				predicted_3d_pos[1, :, :, 0] *= -1
-				predicted_3d_pos[1, :, joints_left + joints_right] = predicted_3d_pos[1, :, joints_right + joints_left]
-				predicted_3d_pos = torch.mean(predicted_3d_pos, dim=0, keepdim=True)
+            # Test-time augmentation (if enabled)
+            if test_generator.augment_enabled():
+                # Undo flipping and take average with non-flipped version
+                predicted_3d_pos[1, :, :, 0] *= -1
+                predicted_3d_pos[1, :, joints_left + joints_right] = predicted_3d_pos[1, :, joints_right + joints_left]
+                predicted_3d_pos = torch.mean(predicted_3d_pos, dim=0, keepdim=True)
+                
+            if return_predictions:
+                return predicted_3d_pos.squeeze(0).cpu().numpy()
+                
+            inputs_3d = torch.from_numpy(batch.astype('float32'))
+            if torch.cuda.is_available():
+                inputs_3d = inputs_3d.cuda()
+            inputs_3d[:, :, 0] = 0    
+            if test_generator.augment_enabled():
+                inputs_3d = inputs_3d[:1]
 
-			if return_predictions:
-				return predicted_3d_pos.squeeze(0).cpu().numpy()
+            error = mpjpe(predicted_3d_pos, inputs_3d)
+            epoch_loss_3d_pos_scale += inputs_3d.shape[0]*inputs_3d.shape[1] * n_mpjpe(predicted_3d_pos, inputs_3d).item()
+
+            epoch_loss_3d_pos += inputs_3d.shape[0]*inputs_3d.shape[1] * error.item()
+            N += inputs_3d.shape[0] * inputs_3d.shape[1]
+            
+            inputs = inputs_3d.cpu().numpy().reshape(-1, inputs_3d.shape[-2], inputs_3d.shape[-1])
+            predicted_3d_pos = predicted_3d_pos.cpu().numpy().reshape(-1, inputs_3d.shape[-2], inputs_3d.shape[-1])
+
+            epoch_loss_3d_pos_procrustes += inputs_3d.shape[0]*inputs_3d.shape[1] * p_mpjpe(predicted_3d_pos, inputs)
+
+            # Compute velocity error
+            epoch_loss_3d_vel += inputs_3d.shape[0]*inputs_3d.shape[1] * mean_velocity_error(predicted_3d_pos, inputs)
+            
+    if action is None:
+        print('----------')
+    else:
+        print('----'+action+'----')
+    e1 = (epoch_loss_3d_pos / N)*1000
+    e2 = (epoch_loss_3d_pos_procrustes / N)*1000
+    e3 = (epoch_loss_3d_pos_scale / N)*1000
+    ev = (epoch_loss_3d_vel / N)*1000
+    print('Test time augmentation:', test_generator.augment_enabled())
+    print('Protocol #1 Error (MPJPE):', e1, 'mm')
+    print('Protocol #2 Error (P-MPJPE):', e2, 'mm')
+    print('Protocol #3 Error (N-MPJPE):', e3, 'mm')
+    print('Velocity Error (MPJVE):', ev, 'mm')
+    print('----------')
+
+    return e1, e2, e3, ev
 
 
 if args.render:
-	print('Rendering...')
+    print('Rendering...')
+    my_action = 'Directions'
+    #input_keypoints = keypoints[args.viz_subject][args.viz_action][args.viz_camera].copy()
+    input_keypoints = keypoints[args.viz_subject][my_action][args.viz_camera].copy()
+    if args.viz_subject in dataset.subjects() and args.viz_action in dataset[args.viz_subject]:
+        ground_truth = dataset[args.viz_subject][args.viz_action]['positions_3d'][args.viz_camera].copy()
+        ##Changes
+        ground_truth = None
+    else:
+        ground_truth = None
+        print('INFO: this action is unlabeled. Ground truth will not be rendered.')
+        
+    gen = UnchunkedGenerator(None, None, [input_keypoints],
+                             pad=pad, causal_shift=causal_shift, augment=args.test_time_augmentation,
+                             kps_left=kps_left, kps_right=kps_right, joints_left=joints_left, joints_right=joints_right)
+    prediction = evaluate(gen, return_predictions=True)
+    
+    if ground_truth is not None:
+        # Reapply trajectory
+        trajectory = ground_truth[:, :1]
+        ground_truth[:, 1:] += trajectory
+        prediction += trajectory
+    
+    # Invert camera transformation
+    cam = dataset.cameras()[args.viz_subject][args.viz_camera]
+    if ground_truth is not None:
+        prediction = camera_to_world(prediction, R=cam['orientation'], t=cam['translation'])
+        ground_truth = camera_to_world(ground_truth, R=cam['orientation'], t=cam['translation'])
+    else:
+        # If the ground truth is not available, take the camera extrinsic params from a random subject.
+        # They are almost the same, and anyway, we only need this for visualization purposes.
+        for subject in dataset.cameras():
+            if 'orientation' in dataset.cameras()[subject][args.viz_camera]:
+                rot = dataset.cameras()[subject][args.viz_camera]['orientation']
+                break
+        prediction = camera_to_world(prediction, R=rot, t=0)
+        # We don't have the trajectory, but at least we can rebase the height
+        prediction[:, :, 2] -= np.min(prediction[:, :, 2])
+    
+    anim_output = {'Reconstruction': prediction}
+    if ground_truth is not None and not args.viz_no_ground_truth:
+        anim_output['Ground truth'] = ground_truth
+    
+    input_keypoints = image_coordinates(input_keypoints[..., :2], w=width_of, h=height_of)
 
-	input_keypoints = keypoints[args.viz_subject][args.viz_action][args.viz_camera].copy()
+    manual_fps = 25
 
-	ground_truth = None
-	print('INFO: this action is unlabeled. Ground truth will not be rendered.')
+    from common.visualization import render_animation
+    render_animation(input_keypoints, anim_output,
+                     dataset.skeleton(), manual_fps, args.viz_bitrate, cam['azimuth'], args.viz_output,
+                     limit=args.viz_limit, downsample=args.viz_downsample, size=args.viz_size,
+                     input_video_path=args.viz_video, viewport=(cam['res_w'], cam['res_h']),
+                     input_video_skip=args.viz_skip)
+    
+else:
+    print('Evaluating...')
+    all_actions = {}
+    all_actions_by_subject = {}
+    for subject in subjects_test:
+        if subject not in all_actions_by_subject:
+            all_actions_by_subject[subject] = {}
 
-	gen = UnchunkedGenerator(None, None, [input_keypoints],
-							 pad=pad, causal_shift=causal_shift, augment=args.test_time_augmentation,
-							 kps_left=kps_left, kps_right=kps_right, joints_left=joints_left, joints_right=joints_right)
+        for action in dataset[subject].keys():
+            action_name = action.split(' ')[0]
+            if action_name not in all_actions:
+                all_actions[action_name] = []
+            if action_name not in all_actions_by_subject:
+                all_actions_by_subject[subject][action_name] = []
+            all_actions[action_name].append((subject, action))
+            all_actions_by_subject[subject][action_name].append((subject, action))
 
-	prediction = evaluate(gen, return_predictions=True)
-	rot = np.asarray([ 0.14070565, -0.15007018, -0.7552408, 0.62232804]).astype(np.float32)
-	prediction = camera_to_world(prediction, R=rot, t=0)
-	# We don't have the trajectory, but at least we can rebase the height
-	prediction[:, :, 2] -= np.min(prediction[:, :, 2])
+    def fetch_actions(actions):
+        out_poses_3d = []
+        out_poses_2d = []
 
-	anim_output = {'Reconstruction': prediction}
-	if ground_truth is not None and not args.viz_no_ground_truth:
-		anim_output['Ground truth'] = ground_truth
+        for subject, action in actions:
+            poses_2d = keypoints[subject][action]
+            for i in range(len(poses_2d)): # Iterate across cameras
+                out_poses_2d.append(poses_2d[i])
 
-	input_keypoints = image_coordinates(input_keypoints[..., :2], w=960, h=720)
+            poses_3d = dataset[subject][action]['positions_3d']
+            assert len(poses_3d) == len(poses_2d), 'Camera count mismatch'
+            for i in range(len(poses_3d)): # Iterate across cameras
+                out_poses_3d.append(poses_3d[i])
 
-	from common.visualization import render_animation
+        stride = args.downsample
+        if stride > 1:
+            # Downsample as requested
+            for i in range(len(out_poses_2d)):
+                out_poses_2d[i] = out_poses_2d[i][::stride]
+                if out_poses_3d is not None:
+                    out_poses_3d[i] = out_poses_3d[i][::stride]
+        
+        return out_poses_3d, out_poses_2d
 
-	human36m_skeleton = Skeleton(parents=[-1,  0,  1,  2,  0, 4,  5,  0,  7,  8,  9,  8, 11, 12, 8, 14, 15],
-	   joints_left=[4, 5,6 ,11, 12,13],
-	   joints_right=[1,2,3,14,15,16])
+    def run_evaluation(actions, action_filter=None):
+        errors_p1 = []
+        errors_p2 = []
+        errors_p3 = []
+        errors_vel = []
 
-	render_animation(input_keypoints, anim_output,
-					 human36m_skeleton, 10, args.viz_bitrate, 70, args.viz_output,
-					 limit=args.viz_limit, downsample=args.viz_downsample, size=args.viz_size,
-					 input_video_path=args.viz_video, viewport=(960, 720),
-					 input_video_skip=args.viz_skip)
+        for action_key in actions.keys():
+            if action_filter is not None:
+                found = False
+                for a in action_filter:
+                    if action_key.startswith(a):
+                        found = True
+                        break
+                if not found:
+                    continue
+
+            poses_act, poses_2d_act = fetch_actions(actions[action_key])
+            gen = UnchunkedGenerator(None, poses_act, poses_2d_act,
+                                     pad=pad, causal_shift=causal_shift, augment=args.test_time_augmentation,
+                                     kps_left=kps_left, kps_right=kps_right, joints_left=joints_left, joints_right=joints_right)
+            e1, e2, e3, ev = evaluate(gen, action_key)
+            errors_p1.append(e1)
+            errors_p2.append(e2)
+            errors_p3.append(e3)
+            errors_vel.append(ev)
+
+        print('Protocol #1   (MPJPE) action-wise average:', round(np.mean(errors_p1), 1), 'mm')
+        print('Protocol #2 (P-MPJPE) action-wise average:', round(np.mean(errors_p2), 1), 'mm')
+        print('Protocol #3 (N-MPJPE) action-wise average:', round(np.mean(errors_p3), 1), 'mm')
+        print('Velocity      (MPJVE) action-wise average:', round(np.mean(errors_vel), 2), 'mm')
+
+    if not args.by_subject:
+        run_evaluation(all_actions, action_filter)
+    else:
+        for subject in all_actions_by_subject.keys():
+            print('Evaluating on subject', subject)
+            run_evaluation(all_actions_by_subject[subject], action_filter)
+            print('')
